+++ packages/app/features/session/lib/src/data/providers/bookmarked_sessions.freezed.dart
