+++ packages/app/features/session/lib/src/data/providers/session_timeline.dart
@@ -1,10 +1,8 @@
+import 'package:app_features_session/src/data/model/timeline_item.dart';
 import 'package:common_data/session.dart';
-import 'package:common_data/sponsor.dart';
-import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:riverpod/riverpod.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
-part 'session_timeline.freezed.dart';
 part 'session_timeline.g.dart';
 
 @riverpod
@@ -40,25 +38,4 @@ Future<List<TimelineItem>> sessionTimeline(Ref ref) async {
   }
 
   return timelineItems;
-}
-
-@freezed
-class TimelineItem with _$TimelineItem {
-  const factory TimelineItem.event({
-    required String title,
-    required DateTime startsAt,
-    required DateTime endsAt,
-  }) = TimelineItemEvent;
-
-  const factory TimelineItem.session({
-    required String id,
-    required String title,
-    required String description,
-    required DateTime startsAt,
-    required DateTime endsAt,
-    required bool isLightningTalk,
-    required SessionVenue venue,
-    required List<ProfileWithSns> speakers,
-    required List<Sponsor> sponsors,
-  }) = TimelineItemSession;
 }
