+++ packages/app/features/session/lib/src/ui/bookmarked_sessions_page.dart
@@ -1,6 +1,7 @@
 import 'package:app_features_session/l10n.dart';
-import 'package:app_features_session/src/providers/bookmarked_sessions.dart';
-import 'package:app_features_session/src/providers/session_timeline.dart';
+import 'package:app_features_session/src/data/model/timeline_item.dart';
+import 'package:app_features_session/src/data/providers/bookmarked_sessions.dart';
+import 'package:app_features_session/src/data/providers/session_timeline.dart';
 import 'package:app_features_session/src/routing/router.dart';
 import 'package:app_features_session/src/ui/session_card.dart';
 import 'package:flutter/material.dart';
